@@ -1,7 +1,7 @@
-const globals = require('globals');
-const js = require('@eslint/js');
+import globals from 'globals';
+import js from '@eslint/js';
 
-module.exports = [
+export default [
 	js.configs.recommended,
 
 	{
