@@ -1,6 +1,6 @@
-const { Node } = require('linkedom');
+import { Node } from 'linkedom';
 
-module.exports = function(window) {
+export default function(window) {
 	const content = window.document.getElementById('article-content');
 
 	if (!content) return;
